feat: finished???
i cannot for the life of me make a responsive design.
Hopefully someday i'll be able to create a site that looks good on cellphone aswell. This anyway is kind of nice and i learned a lot of useful things. Maybe i'll try to add that clock from before somewhere here. But i'm done for the time being